ch(refactor) refactor project structure
[finishes #167276572]
